requirements.txt
# Core Google API libraries
google-api-python-client>=2.100.0
google-auth>=2.20.0

# Claude API
anthropic>=0.40.0
